#696 Creates IndexSearchbar
