+++ src/components/index.ts
@@ -11,6 +11,7 @@ export * from './hamburgerMenu'
 export * from './heading'
 export * from './htmlParser'
 export * from './icon'
+export * from './indexSearchbar'
 export * from './input'
 export * from './inputGroup'
 export * from './label'
@@ -19,7 +20,6 @@ export * from './modal'
 export * from './paginationWrapper'
 export * from './publishGameForm'
 export * from './richTextEditor'
-export * from './searchbar'
 export * from './spinner'
 export * from './svgs'
 export * from './tab'
