+++ src/components/searchbar/__tests__/searchbar.test.tsx
@@ -1,10 +1,5 @@
 import React from 'react'
-import { fireEvent, render, screen, waitFor } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
-import mockRouter from 'next-router-mock'
-
-import { ROUTES } from '@/constants'
-import { sleep } from '@/test/helpers'
+import { render, screen } from '@testing-library/react'
 
 import { Searchbar } from '..'
 
@@ -43,62 +38,20 @@ describe('<Searchbar />', () => {
     })
   })
 
-  describe('searching', () => {
-    it("should not update the route with the user's query automatically after 1 seconds when autoSearch is false", async () => {
-      mockRouter.push(ROUTES.adminGames)
-      render(<Searchbar />)
-      userEvent.type(screen.getByTestId('searchbar'), 'query')
-      await waitFor(
-        async () => {
-          await sleep(1000)
-          expect(mockRouter.asPath).toBe(ROUTES.adminGames)
-        },
-        { timeout: 1500 },
-      )
+  describe('clear button', () => {
+    it('should render a clear button when showingClearButton is true', () => {
+      render(<Searchbar showingClearButton />)
+      expect(screen.getByTestId('searchbar-clear-button')).toBeVisible()
     })
 
-    it("should update the route with the user's query automatically after 1 seconds when autoSearch is true", async () => {
-      mockRouter.push(ROUTES.adminGames)
-      render(<Searchbar autoSearch />)
-      userEvent.type(screen.getByTestId('searchbar'), 'query')
-      await waitFor(
-        () => {
-          expect(mockRouter.asPath).toBe(`${ROUTES.adminGames}?query=query`)
-        },
-        { timeout: 1500 },
+    it('should pass clear button props', () => {
+      render(
+        <Searchbar
+          clearButtonProps={{ className: 'mr-4' }}
+          showingClearButton
+        />,
       )
-    })
-
-    it('should allow user to enter the search with the enter key', async () => {
-      mockRouter.push(ROUTES.adminGames)
-      render(<Searchbar />)
-      await userEvent.type(screen.getByTestId('searchbar'), 'query')
-      fireEvent.keyUp(screen.getByTestId('searchbar'), {
-        currentTarget: { value: 'query' },
-        key: 'Enter',
-      })
-      expect(mockRouter.asPath).toBe(`${ROUTES.adminGames}?query=query`)
-    })
-
-    it('should render a button to clear the search', async () => {
-      mockRouter.push(`${ROUTES.adminGames}?platforms[]=1&query=query`)
-      render(<Searchbar />)
-      userEvent.click(screen.getByLabelText('Clear Search'))
-      await waitFor(() =>
-        expect(mockRouter.asPath).toBe(
-          encodeURI(`${ROUTES.adminGames}?platforms[]=1`),
-        ),
-      )
-    })
-
-    it('should use the given pathname', async () => {
-      mockRouter.push(ROUTES.home)
-      render(<Searchbar pathnameOverride={ROUTES.games} />)
-      await userEvent.type(screen.getByTestId('searchbar'), 'query')
-      await userEvent.type(screen.getByTestId('searchbar'), '{enter}')
-      await waitFor(() =>
-        expect(mockRouter.asPath).toBe(`${ROUTES.games}?query=query`),
-      )
+      expect(screen.getByTestId('searchbar-clear-button')).toHaveClass('mr-4')
     })
   })
 })
