+++ src/components/searchbar/searchbar.tsx
@@ -1,22 +1,73 @@
-import React, { Suspense } from 'react'
+import React, { HTMLAttributes } from 'react'
+import clsx from 'clsx'
 
+import { Icon } from '../icon'
 import { Input } from '../input'
-import { SearchbarClientComponent } from './clientComponent'
 
-export function Searchbar(props: PropsOf<typeof SearchbarClientComponent>) {
+interface Props extends HTMLAttributes<HTMLDivElement> {
+  clearButtonProps?: HTMLAttributes<HTMLButtonElement>
+  inputProps?: Omit<PropsOf<typeof Input>, 'id'>
+  labelProps?: Omit<HTMLAttributes<HTMLLabelElement>, 'htmlFor'> & {
+    ariaLabel?: string
+  }
+  showingClearButton?: boolean
+}
+
+export function Searchbar({
+  className,
+  clearButtonProps = {},
+  id = 'searchbar',
+  inputProps = {},
+  labelProps = { ariaLabel: 'Search by Name' },
+  showingClearButton,
+  ...options
+}: Props) {
+  const {
+    ariaLabel = 'Search by Name',
+    className: labelClassName,
+    ...labelOptions
+  } = labelProps
+  const { className: inputClassName, ...inputOptions } = inputProps
+  const { className: clearButtonClassName, ...clearButtonOptions } =
+    clearButtonProps
+
   return (
-    <Suspense
-      fallback={
-        <Input
-          autoComplete="off"
-          className="skeleton px-8"
-          data-testid={props.id}
-          id={props.id}
-          placeholder="Search"
-        />
-      }
-    >
-      <SearchbarClientComponent {...props} />
-    </Suspense>
+    <div className={clsx('relative', className)} {...options}>
+      <label
+        aria-label={ariaLabel}
+        className={clsx(
+          'absolute top-1/2 left-2 -translate-y-1/2 transform cursor-pointer',
+          'text-heading-md text-neutral-500',
+          labelClassName,
+        )}
+        htmlFor={id}
+        {...labelOptions}
+      >
+        <Icon icon="magnify" />
+      </label>
+      <Input
+        autoComplete="off"
+        className={clsx('px-8', inputClassName)}
+        data-testid={id}
+        id={id}
+        placeholder="Search"
+        {...inputOptions}
+      />
+      {showingClearButton ? (
+        <button
+          aria-label="Clear Search"
+          className={clsx(
+            'absolute top-[9px] right-2 cursor-pointer text-2xl',
+            'text-neutral-500 hover:opacity-70',
+            clearButtonClassName,
+          )}
+          data-testid="searchbar-clear-button"
+          type="button"
+          {...clearButtonOptions}
+        >
+          <Icon icon="close" />
+        </button>
+      ) : null}
+    </div>
   )
 }
